refactor: move all src types to types.d.ts
